reset password

+++ src/controllers/UserController.js
@@ -125,12 +125,20 @@ class UserController {
     // [POST] /user/check-email
     async checkEmail(req, res, next) {
         try {
-            const { email } = req.body
+            const { email, mode } = req.body
 
-            // Kiểm tra email đã tồn tại chưa
-            const existingUser = await User.findOne({ email, password: { $ne: '' }, id: '' })
-            if (existingUser) {
-                return res.status(400).json({ message: 'Email đã tồn tại' })
+            if (mode === 'signup') {
+                // Kiểm tra email đã tồn tại chưa
+                const existingUser = await User.findOne({ email, password: { $ne: '' }, id: '' })
+                if (existingUser) {
+                    return res.status(400).json({ message: 'Email đã tồn tại' })
+                }
+            } else if (mode === 'forgot-password') {
+                // Kiểm tra email đã tồn tại chưa
+                const existingUser = await User.findOne({ email, password: { $ne: '' }, id: '' })
+                if (!existingUser) {
+                    return res.status(400).json({ message: 'Email không tồn tại' })
+                }
             }
 
             // Tạo mã xác thực ngẫu nhiên 6 số
@@ -154,7 +162,7 @@ class UserController {
     // [POST] /user/verify-email
     async verifyEmail(req, res, next) {
         try {
-            const { email, code, password } = req.body
+            const { email, code, password, mode } = req.body
 
             // Kiểm tra mã xác thực
             const verification = await VerificationCode.findOne({ email, code })
@@ -162,14 +170,39 @@ class UserController {
                 return res.status(400).json({ message: 'Mã xác thực không chính xác hoặc đã hết hạn' })
             }
 
-            // Tạo tài khoản mới
-            const user = new User({ email, password })
+            if (mode === 'signup') {
+                // Tạo tài khoản mới
+                const user = new User({ email, password })
+                await user.save()
+                // Xóa mã xác thực
+                await VerificationCode.deleteOne({ email, code })
+
+                res.status(201).json({ user: { id: user._id, email: user.email } })
+            } else if (mode === 'forgot-password') {
+                // Xóa mã xác thực
+                res.status(200).json({ message: 'Mã xác thực đã được xác nhận' })
+            }
+        } catch (err) {
+            next(err)
+        }
+    }
+
+    // [POST] /user/reset-password
+    async resetPassword(req, res, next) {
+        try {
+            const { email, code, password } = req.body
+            const user = await User.findOne({ email, password: { $ne: '' }, id: '' })
+            if (!user) {
+                return res.status(400).json({ message: 'Email không tồn tại' })
+            }
+            const verification = await VerificationCode.findOne({ email, code })
+            if (!verification) {
+                return res.status(400).json({ message: 'Mã xác thực không chính xác hoặc đã hết hạn' })
+            }
+            user.password = password
             await user.save()
-
-            // Xóa mã xác thực
             await VerificationCode.deleteOne({ email, code })
-
-            res.status(201).json({ user: { id: user._id, email: user.email } })
+            return res.status(200).json({ message: 'Mật khẩu đã được đặt lại thành công' })
         } catch (err) {
             next(err)
         }
